Backend/requirements.txt
djangorestframework
django
google-api-python-client
pillow
